+++ src/app/components/layouts/signin-layout/signin-layout.component.ts
@@ -1,11 +1,12 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { RouterModule } from '@angular/router';
 
-import { SigninComponent } from '@components/user/signin/signin.component';
+import { UnsignedToolbarComponent } from '@components/navigation/unsigned-toolbar/unsigned-toolbar.component';
 
 @Component({
   selector: 'ltm-signin-layout',
   standalone: true,
-  imports: [SigninComponent],
+  imports: [RouterModule, UnsignedToolbarComponent],
   templateUrl: './signin-layout.component.html',
   styleUrl: './signin-layout.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,