+++ src/app/services/theme/theme.service.ts
@@ -1,16 +1,19 @@
 import { Injectable, WritableSignal, inject, signal } from '@angular/core';
 import { OverlayContainer } from '@angular/cdk/overlay';
 
+import { AuthService } from '@services/auth/auth.service';
 import {
   theme as configTheme,
   availableThemes as configAvailableThemes,
 } from '@config';
+import { Observable, of } from 'rxjs';
 
 @Injectable({
   providedIn: 'root',
 })
 export class ThemeService {
   private readonly overlayContainer = inject(OverlayContainer);
+  private readonly auth = inject(AuthService);
 
   readonly #$theme: WritableSignal<string>;
   public readonly $theme;
@@ -18,13 +21,13 @@ export class ThemeService {
 
   /**
    * Defines the theme to use.
-   * First it checks in localStorage, then in the config file, and if it all
+   * First it checks in user prefs, then in the config file, and if it all
    * fails then it tries to get the first one in availableThemes.
    * If everything fails, it sets an empty theme.
    */
   constructor() {
     let newTheme: string;
-    const storedTheme = localStorage.getItem('LT_theme');
+    const storedTheme = this.auth.getPref('theme');
 
     if (storedTheme && this.availableThemes.includes(storedTheme)) {
       newTheme = storedTheme;
@@ -40,13 +43,13 @@ export class ThemeService {
     if (newTheme) this.setOverlayTheme(newTheme);
   }
 
-  switchTheme(newTheme: string) {
+  switchTheme(newTheme: string): Observable<any> {
     if (this.availableThemes.includes(newTheme)) {
       this.setOverlayTheme(newTheme, this.#$theme());
       this.#$theme.set(newTheme);
 
-      localStorage.setItem('LT_theme', newTheme);
-    }
+      return this.auth.setPref({ theme: newTheme });
+    } else return of(null);
   }
 
   setOverlayTheme(newTheme: string, prevTheme?: string): void {